Pass under Miri
The main change is for `alloc` to push new slabs onto the `Vec` before
forming the raw `next` pointer. Under Stacked Borrows, moving the `Box`
includes a retag operation, which was invalidating `next`.

A secondary change is to remove int-to-pointer casts (used for alignment).

The `alloc` functions are refactored to reduce duplication and move the
whole slow path out of `alloc` itself.

Finally, the doctest's node initialization now avoids forming references
to uninitialized memory.